+++ chaosreport/__init__.py
@@ -10,7 +10,7 @@ import subprocess
 import tempfile
 from base64 import b64encode
 from datetime import datetime, timedelta
-from typing import List
+from typing import Any, Dict, List
 
 import cairosvg
 import dateparser
@@ -20,8 +20,11 @@ import maya
 import pygal
 import pypandoc
 import semver
+from chaoslib import merge_vars, substitute
 from chaoslib.caching import cache_activities, lookup_activity
-from chaoslib.types import Experiment, Journal, Run
+from chaoslib.configuration import load_configuration
+from chaoslib.secret import load_secrets
+from chaoslib.types import Configuration, Experiment, Journal, Run, Secrets
 from jinja2 import Environment, PackageLoader
 from logzero import logger
 from natural import date
@@ -244,7 +247,12 @@ def generate_report_header(
     return header
 
 
-def generate_report(journal_path: str, export_format: str = "markdown") -> str:
+def generate_report(
+    journal_path: str,
+    export_format: str = "markdown",
+    var: Dict[str, Any] = None,
+    var_file: List[str] = None,
+) -> str:
     """
     Generate a report document from a chaostoolkit journal.
 
@@ -267,7 +275,14 @@ def generate_report(journal_path: str, export_format: str = "markdown") -> str:
 
     generate_chart_from_metric_probes(journal, export_format)
     add_contribution_model(journal, export_format)
-    template = get_report_template(journal["chaoslib-version"])
+    config_vars, secret_vars = merge_vars(var, var_file) or (None, None)
+    config = load_configuration(
+        experiment.get("configuration", {}), config_vars
+    )
+    secrets = load_secrets(experiment.get("secrets", {}), secret_vars)
+    template = get_report_template(
+        journal["chaoslib-version"], configuration=config, secrets=secrets
+    )
     report = template.render(journal)
 
     return report
@@ -333,7 +348,10 @@ def save_report(
 
 
 def get_report_template(
-    report_version: str, default_template: str = "index.md"
+    report_version: str,
+    default_template: str = "index.md",
+    configuration: Configuration = None,
+    secrets: Secrets = None,
 ):
     """
     Retrieve and return the most appropriate template based on the
@@ -346,6 +364,13 @@ def get_report_template(
     env.globals["pretty_duration"] = lambda d0, d1: date.delta(
         dateparser.parse(d0), dateparser.parse(d1), words=False
     )[0]
+    env.globals["substitute"] = (
+        lambda args: {
+            k: substitute(v, configuration, secrets) for k, v in args.items()
+        }
+        if isinstance(args, dict)
+        else args
+    )
 
     if not report_version:
         return env.get_template(default_template)